mongodb test infrastructure. Test for post session
