--- e2e/db.js
@@ -1,13 +1,0 @@
-const mongoose = require('mongoose');
-
-module.exports = {   
-  dropCollection(name) {
-    return mongoose.connection.dropCollection(name)
-      .catch(err => {
-        if(err.codeName !== 'NamespaceNotFound') throw err;
-      });
-  },
-  dropDatabase() {
-    return mongoose.connection.dropDatabase();
-  }
-};
