+++ e2e/db.ts
@@ -1,0 +1,10 @@
+import mongoose from 'mongoose';
+
+export const dropCollection = (name: string) => {
+  return mongoose.connection.dropCollection(name)
+    .catch( (err: mongoose.Error) => {
+      if(err.codeName !== 'NamespaceNotFound') throw err;
+    });
+};
+
+export const dropDatabase = () => mongoose.connection.dropDatabase();
